添加mac对openssl和mysql的支持

+++ src/Util/SSLBox.cpp
@@ -41,7 +41,7 @@ SSL_Initor::SSL_Initor() {
 		_mutexes[n].unlock();
 	});
 	CRYPTO_set_id_callback([]() ->unsigned long {
-		return pthread_self();
+		return (unsigned long)pthread_self();
 	});
 	ssl_client = SSL_CTX_new(TLSv1_client_method());
 	ssl_server = SSL_CTX_new(TLSv1_server_method());
